# [FIX] sale_coupon: various performance update