Various update in the code in order to decrease the execution time of the computation of the sale coupon.
Tested using the client database (~2000 coupon programs).
The time is measured on the TTFB of the shoping cart opening with the server restarted
(other test were made by updating the cart, but are not relevant to point out here):
 - Before this commit: 3 s
 - With this commit: 1.7 s
(Note that it introduce the rendering of the webpage as well which is expensive but cached after first attempt)

With the cache, updating the cart execution time TTFB:
 - Before this commit: 2.7 s
 - With this commit: 0.8 s

pyflame graphs are available in the ticket attachment

Here is a more detailed of the different modication that were made:
 - In `_get_valid_applied_coupon_program` we did call for all applicable programs.
 However only the ones which are on the existing order are usefull to check if they are still applicable.
 Result: 3s -> 2.1s
 - Use set to store program ids in `_filter_on_mimimum_amount`
 Result: 2.1s -> 1.9s
 - Avoid call `_get_valid_products` in `_get_valid_products`
 Result: 1.9s -> 1.7s

OPW-2323599

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr